refactor: env/ 경로 생성
process.env 처리 위치를 통일하기 위한 경로입니다.
