+++ .gitignore
@@ -74,6 +74,10 @@ typings/
 .env*
 .env.*
 
+# type safe env parsing
+!backend/src/env/
+!backend/src/env/*.ts
+
 # parcel-bundler cache (https://parceljs.org/)
 .cache
